Added the search and add mem capabilities

+++ topic3-2/src/main/java/com/josh/demo/controllers/AddMemController.java
@@ -28,7 +28,7 @@ public class AddMemController
 	@PostMapping("/add")
 	public String addMem(MemModel mem, Model model)
 	{
-		int result = ordersService.addOne(mem);
+		long result = ordersService.addOne(mem);
 		if(result > 0){
 			model.addAttribute("memModel2", "Model added! " + mem.toString());
 		}
